movning teamname cmds to sport
