+++ lang/mod.json
@@ -17,11 +17,6 @@
         "help_plugin_load": "Load a plugin",
         "help_plugin_reload": "Reload a plugin",
         "help_plugin_available": "List available plugins",
-        "help_teamname": "Manage teamnames of sport teams",
-        "help_teamname_info": "Show name variants for a team",
-        "help_teamname_set": "Update name variant for a team",
-        "help_teamname_add": "Add a new team",
-        "help_teamname_del": "Remove an entry",
 
         "#1": "help descs",
         "desc_about": "Print the credits.",
@@ -69,14 +64,6 @@
         "errors_on_reload": "Error on reloading plugin `{}`.",
         "plugins_avail": "**{} additional plugins available**",
         "no_plugin_avail": "All available plugins already loaded. There are no more unloaded plugins available.",
-        "team_not_found": "Team not found.",
-        "teamname_long": "**Long**",
-        "teamname_short": "**Short**",
-        "teamname_abbr": "**Abbr.**",
-        "teamname_other": "Other",
-        "teamname_set_variant_invalid": "Variant must be one of the following: {}",
-        "teamname_set_duplicate": "`{}` is already existent for the team of {}.",
-        "teamname_added": "Names added. ({})",
 
         "#3": "ignore",
         "nothing_blocked": "No users or commands blocked.",
@@ -127,11 +114,6 @@
         "help_plugin_load": "Lädt ein Plugin",
         "help_plugin_reload": "Lädt ein Plugin neu",
         "help_plugin_available": "Listet alle verfügbaren Plugins",
-        "help_teamname": "Verwaltet Namensvarianten von Sportvereinen",
-        "help_teamname_info": "Zeigt alle Varianten zu einem Team an",
-        "help_teamname_set": "Ändert eine Namensvariante",
-        "help_teamname_add": "Fügt Namensvarianten für einen neuen Verein hinzu",
-        "help_teamname_del": "Entfernt einen Verein",
 
         "#1": "help descs",
         "desc_about": "Zeigt die Credits an.",
@@ -179,14 +161,6 @@
         "errors_on_reload": "Fehler beim Neuladen vom Plugin `{}`.",
         "plugins_avail": "**{} weitere Plugins verfügbar**",
         "no_plugin_avail": "Alle verfügbaren Plugins sind bereits geladen. Es sind keine weiteren Plugins verfügbar.",
-        "team_not_found": "Team wurde nicht gefunden.",
-        "teamname_long": "**Lang**",
-        "teamname_short": "**Kurz**",
-        "teamname_abbr": "**Abk.**",
-        "teamname_other": "Weitere",
-        "teamname_set_variant_invalid": "Variante muss eine der folgenden sein: {}",
-        "teamname_set_duplicate": "`{}` ist bereits für einen anderen Verein ({}) eingespeichert.",
-        "teamname_added": "Namen zum Verein ({}) hinzugefügt",
 
         "#3": "ignore",
         "nothing_blocked": "Keine User oder Kommandos blockiert.",
